whmcs-2.0.6

+++ modules/addons/SSLCENTERWHMCS/models/apiConfiguration/Repository.php
@@ -213,6 +213,9 @@ class Repository extends \MGModule\SSLCENTERWHMCS\mgLibs\models\Repository
 
     public function dropApiConfigurationTable()
     {
-        Capsule::schema()->dropIfExists($this->tableName);
+        if (Capsule::schema()->hasTable($this->tableName))
+        {
+            Capsule::schema()->dropIfExists($this->tableName);
+        }
     }
 }
